User Patient schema updated. (#25)
