+++ android-bypass/config-android.ts
@@ -1,7 +1,8 @@
 export const MOCK_USER = {
     id: '116238727817152771287',
     email: "dev.mock.user@email.com",
-    name: "Dev Mock"
+    name: "Dev Mock",
+    picture: "https://ui-avatars.com/api/?name=Dev+Mock&background=49AFBE&color=FEFEFF"
 };
 
 export const FORCE_DB_RESET_ANDROID = false;
