+++ android-bypass/google-auth-android.ts
@@ -1,11 +1,11 @@
 // --------- Mock Session for Android
-import { Platform } from "react-native";
 import { getUserFromStorage, saveUser, signOut } from "@/services/auth-service/google-auth";
-import * as SecureStore from "expo-secure-store";
-import { router } from "expo-router";
+import { User } from "@/services/database/migrations/v1/schema_v1";
 import { logger } from "@/services/logging/logger";
 import { ROUTES } from "@/utils/route";
-import { User } from "@/services/database/migrations/v1/schema_v1";
+import { router } from "expo-router";
+import * as SecureStore from "expo-secure-store";
+import { Platform } from "react-native";
 import { MOCK_USER } from "./config-android";
 
 export const isAndroid = Platform.OS === "android";
@@ -13,10 +13,10 @@ export const isAndroid = Platform.OS === "android";
 export const handleMockSignIn = async () => {
     console.log("🤖 Android emulator detected — skipping SSO and using mock user");
 
-    await saveUser(MOCK_USER);
-    const keys = ["access_token", "refresh_token", "issued_at", "expires_in"];
+    const keys = ["access_token", "refresh_token", "issued_at", "expires_in", "user"];
     await Promise.all(keys.map((key) => SecureStore.deleteItemAsync(key)));
     logger.debug("Android - Tokens cleared.");
+    await saveUser(MOCK_USER);
     router.replace(ROUTES.MY_HEALTH);
 };
