+++ services/core/PatientAllergyService.ts
@@ -1,7 +1,8 @@
 import { PatientAllergy } from '@/services/database/migrations/v1/schema_v1';
 import { PatientAllergyModel } from '@/services/database/models/PatientAllergyModel';
 import { logger } from '@/services/logging/logger';
-import { useModel, getCurrentTimestamp } from '@/services/core/utils';
+import { getCurrentTimestamp } from '@/services/core/utils';
+import { useModel } from '@/services/database/BaseModel';
 import { isExistingPatientById } from '@/services/core/PatientService';
 
 // Single shared instance of model
