+++ services/core/PatientEquipmentService.ts
@@ -1,7 +1,8 @@
 import { PatientEquipment } from '@/services/database/migrations/v1/schema_v1';
 import { PatientEquipmentModel } from '@/services/database/models/PatientEquipmentModel';
 import { logger } from '@/services/logging/logger';
-import { useModel, getCurrentTimestamp } from '@/services/core/utils';
+import { getCurrentTimestamp } from '@/services/core/utils';
+import { useModel } from '@/services/database/BaseModel';
 import { isExistingPatientById } from '@/services/core/PatientService';
 
 // Single shared instance of model
