+++ services/core/PatientNoteService.ts
@@ -1,7 +1,8 @@
 import { PatientNote } from '@/services/database/migrations/v1/schema_v1';
 import { PatientNoteModel } from '@/services/database/models/PatientNoteModel';
 import { logger } from '@/services/logging/logger';
-import { useModel, getCurrentTimestamp } from '@/services/core/utils';
+import { getCurrentTimestamp } from '@/services/core/utils';
+import { useModel } from '@/services/database/BaseModel';
 import { isExistingPatientById } from '@/services/core/PatientService';
 
 // Single shared instance of model
